requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
flask==2.3.3
pandas==2.1.1
matplotlib==3.7.2
moviepy==1.0.3
gtts==2.3.2
pillow==10.0.1
schedule==1.2.0
python-telegram-bot==20.6
discord.py==2.3.2
firebase-admin==6.2.0
webdriver-manager==4.0.1
